refactor: restructure connect routes, random chat, and settings UI- Moved Random Chat feature under /connect route- Updated all intro/welcome videos accordingly- Added separate Settings action icon for consistent use in components- Restructured folder layout for better clarity and modularity- Fixed React silent error in signup/login flow
