+++ connect-ui/src/components/common/PrivacyPolicyPage.jsx
@@ -1,6 +1,11 @@
 import React, { useEffect } from 'react';
-import { Container, Typography, Box, Divider } from '@/MUI/MuiComponents';
-import StyledText from '@/components/common/StyledText';
+import {
+  Container,
+  Typography,
+  Box,
+  Divider,
+} from '@mui/material';
+import StyledText from './StyledText';
 
 // === Reusable Section Component ===
 function Section({ title, children }) {
@@ -26,7 +31,7 @@ function PrivacyPolicyPage() {
       {/* Page Title */}
       <Box textAlign="start" mb={4}>
         <Typography variant="h4" fontWeight={600} gutterBottom>
-          <StyledText text="Privacy" /> Policy
+          <StyledText text='Privacy' /> Policy
         </Typography>
         <Typography variant="subtitle1" color="text.secondary">
           How Connect collects, uses, and protects your information.
@@ -84,7 +89,7 @@ function PrivacyPolicyPage() {
 
       <Section title="7. Contact Us">
         If you have any questions or concerns, feel free to contact us at{' '}
-        <StyledText text="support@connectapp.com" />.
+        <Box component="span" color="primary.main">support@connectapp.com</Box>.
       </Section>
 
       <Divider sx={{ my: 4 }} />
