--- connect-ui/src/pages/private/chat/ChatMessageInput.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-// Input box to type and send new messages
-
-function ChatMessageInput() {
-  return <div>ChatMessageInput</div>;
-}
-
-export default ChatMessageInput;
