--- connect-ui/src/pages/private/chat/ChatMessageList.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-// Displays the messages exchanged in the current chat
-
-function ChatMessageList() {
-  return <div>ChatMessageList</div>;
-}
-
-export default ChatMessageList;
