+++ connect-ui/src/pages/private/connect/normal/ChatMessageInput.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+// Input box to type and send new messages
+
+function ChatMessageInput() {
+  return <div>ChatMessageInput</div>;
+}
+
+export default ChatMessageInput;
