+++ connect-ui/src/pages/private/connect/normal/ChatMessageList.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+// Displays the messages exchanged in the current chat
+
+function ChatMessageList() {
+  return <div>ChatMessageList</div>;
+}
+
+export default ChatMessageList;
