+++ connect-ui/src/pages/private/connect/normal/mockData.js
@@ -1,0 +1,63 @@
+
+export const chatPartners = [
+  {
+    id: 'u1',
+    name: 'Riya Sharma',
+    status: 'online',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=riya',
+  },
+  {
+    id: 'u2',
+    name: 'Arjun Mehta',
+    status: 'offline',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=arjun',
+  },
+  {
+    id: 'u3',
+    name: 'Sneha Patel',
+    status: 'online',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=sneha',
+  },
+  {
+    id: 'u4',
+    name: 'Karan Kapoor',
+    status: 'offline',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=karan',
+  },
+  {
+    id: 'u5',
+    name: 'Meera Iyer',
+    status: 'online',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=meera',
+  },
+  {
+    id: 'u6',
+    name: 'Dev Singh',
+    status: 'offline',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=dev',
+  },
+  {
+    id: 'u7',
+    name: 'Isha Khan',
+    status: 'online',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=isha',
+  },
+  {
+    id: 'u8',
+    name: 'Rahul Joshi',
+    status: 'offline',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=rahul',
+  },
+  {
+    id: 'u9',
+    name: 'Tanya Desai',
+    status: 'online',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=tanya',
+  },
+  {
+    id: 'u10',
+    name: 'Nikhil Rao',
+    status: 'offline',
+    avatar: 'https://api.dicebear.com/8.x/lorelei/svg?seed=nikhil',
+  },
+];
