+++ connect-ui/src/pages/private/connect/random/RandomChatHeader.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function RandomChatHeader() {
+  return (
+    <div>RandomChatHeader</div>
+  )
+}
+
+export default RandomChatHeader
