+++ connect-ui/src/pages/private/connect/random/components/ConnectButton.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { Button } from '@/MUI/MuiComponents';
+import BoltIcon from '@mui/icons-material/Bolt';
+
+function ConnectButton({ onClick, disabled = false }) {
+    return (
+        <Button
+            variant="outlined"
+            fullWidth
+            startIcon={<BoltIcon sx={{ color: 'success.main' }} />}
+            onClick={onClick}
+            disabled={disabled}
+            sx={{ width: 'fit-content', color: 'text.primary' }}
+        >
+            Connect
+        </Button>
+    );
+}
+
+export default ConnectButton;
