--- connect-ui/src/pages/private/mockData.js
@@ -1,145 +1,0 @@
-export const currentUserId = 'user1';
-
-export const chatPartners = [
-  {
-    id: 'user1',
-    name: 'Tan',
-    avatar: 'https://i.pravatar.cc/150?u=tan',
-    status: 'online'
-  },
-  {
-    id: 'user2',
-    name: 'Aisha',
-    avatar: 'https://i.pravatar.cc/150?u=aisha',
-    status: 'offline'
-  },
-  {
-    id: 'user3',
-    name: 'Ravi',
-    avatar: 'https://i.pravatar.cc/150?u=ravi',
-    status: 'busy'
-  },
-  {
-    id: 'user4',
-    name: 'Mei',
-    avatar: 'https://i.pravatar.cc/150?u=mei',
-    status: 'online'
-  },
-  {
-    id: 'user5',
-    name: 'Liam',
-    avatar: 'https://i.pravatar.cc/150?u=liam',
-    status: 'offline'
-  },
-  {
-    id: 'user6',
-    name: 'Zara',
-    avatar: 'https://i.pravatar.cc/150?u=zara',
-    status: 'online'
-  },
-  {
-    id: 'user7',
-    name: 'Noah',
-    avatar: 'https://i.pravatar.cc/150?u=noah',
-    status: 'busy'
-  },
-  {
-    id: 'user8',
-    name: 'Emma',
-    avatar: 'https://i.pravatar.cc/150?u=emma',
-    status: 'away'
-  },
-  {
-    id: 'user9',
-    name: 'Dev',
-    avatar: 'https://i.pravatar.cc/150?u=dev',
-    status: 'online'
-  },
-  {
-    id: 'user10',
-    name: 'Sofia',
-    avatar: 'https://i.pravatar.cc/150?u=sofia',
-    status: 'offline'
-  },
-  {
-    id: 'user11',
-    name: 'Arjun',
-    avatar: 'https://i.pravatar.cc/150?u=arjun',
-    status: 'online'
-  },
-  {
-    id: 'user12',
-    name: 'Lena',
-    avatar: 'https://i.pravatar.cc/150?u=lena',
-    status: 'offline'
-  },
-  {
-    id: 'user13',
-    name: 'Carlos',
-    avatar: 'https://i.pravatar.cc/150?u=carlos',
-    status: 'busy'
-  },
-  {
-    id: 'user14',
-    name: 'Yuki',
-    avatar: 'https://i.pravatar.cc/150?u=yuki',
-    status: 'online'
-  },
-  {
-    id: 'user15',
-    name: 'Mina',
-    avatar: 'https://i.pravatar.cc/150?u=mina',
-    status: 'away'
-  },
-  {
-    id: 'user16',
-    name: 'Ali',
-    avatar: 'https://i.pravatar.cc/150?u=ali',
-    status: 'offline'
-  },
-  {
-    id: 'user17',
-    name: 'Chloe',
-    avatar: 'https://i.pravatar.cc/150?u=chloe',
-    status: 'online'
-  },
-  {
-    id: 'user18',
-    name: 'Ethan',
-    avatar: 'https://i.pravatar.cc/150?u=ethan',
-    status: 'busy'
-  },
-  {
-    id: 'user19',
-    name: 'Nina',
-    avatar: 'https://i.pravatar.cc/150?u=nina',
-    status: 'online'
-  },
-  {
-    id: 'user20',
-    name: 'Omar',
-    avatar: 'https://i.pravatar.cc/150?u=omar',
-    status: 'offline'
-  }
-];
-
-export const mockMessages = [
-  {
-    id: 'm1',
-    senderId: 'user1',
-    content: 'Hey! How are you?',
-    timestamp: '2025-07-04T12:30:00Z'
-  },
-  {
-    id: 'm2',
-    senderId: 'user2',
-    content: "I'm good! You?",
-    timestamp: '2025-07-04T12:31:00Z'
-  },
-  {
-    id: 'm3',
-    senderId: 'user1',
-    content: 'All good, just testing this chat UI 😊',
-    timestamp: '2025-07-04T12:32:00Z'
-  }
-];
